+++ src/react-components/src/components/Accordion/Accordion.stories.tsx
@@ -4,7 +4,7 @@ import Accordion from "./Accordion";
 import { action } from "@storybook/addon-actions";
 import { text } from "@storybook/addon-knobs";
 import Checkbox from "../Checkbox/Checkbox";
-import UnorderedList from "../List/UnorderedList";
+import UnorderedList from "../UnorderedList/UnorderedList";
 
 export default {
     title: "Accordion",
